requirements.txt
aiofiles==0.6.0
ansicon==1.89.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
blessed==1.18.0
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
ElasticMock==1.8.0
elasticsearch==7.9.1
entrypoints==0.3
fastapi==0.63.0
fitparse==1.2.0
folium==0.11.0
gpxpy==1.4.2
h11==0.12.0
idna==2.10
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jinxed==1.0.1
joblib==1.0.1
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mock==4.0.3
nbformat==5.1.2
nest-asyncio==1.5.1
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.18
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2021.1
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.24.0
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
starlette==0.13.6
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.13.4
wcwidth==0.2.5
webencodings==0.5.1
xgboost==1.3.3
zipp==3.4.1
